+++ common/encoding.go
@@ -400,7 +400,7 @@ func (c *Change) convertProto() *ChangePb {
 		cp.CommitIndex = proto.Uint32(c.CommitIndex)
 	}
 	if c.TransactionID != 0 {
-		cp.TransactionID = proto.Uint32(c.TransactionID)
+		cp.TransactionID = proto.Uint64(c.TransactionID)
 	}
 	cp.NewColumns = unmakeRow(c.NewRow)
 	cp.OldColumns = unmakeRow(c.OldRow)
